+++ activemodel/lib/active_model/validations/model.rb
@@ -1,23 +1,36 @@
 # frozen_string_literal: true
 
-module ActiveRecord
+module ActiveModel
   module Validations
-    class AssociatedValidator < ActiveModel::Validations::ModelValidator # :nodoc:
+    class ModelValidator < EachValidator # :nodoc:
+      def validate_each(record, attribute, value)
+        context = record_validation_context_for_association(record)
+
+        if Array(value).reject { |object| valid_object?(object, context) }.any?
+          record.errors.add(attribute, :invalid, **options.merge(value: value))
+        end
+      end
+
       private
-        def valid_object?(record, context)
-          (record.respond_to?(:marked_for_destruction?) && record.marked_for_destruction?) || record.valid?(context)
+        def valid_object?(object, context)
+          object.valid?(context)
+        end
+
+        def record_validation_context_for_association(record)
+          record.custom_validation_context? ? record.validation_context : nil
         end
     end
 
     module ClassMethods
       # Validates whether the associated object or objects are all valid.
-      # Works with any kind of association.
+      # Each associated object has to repond to the method <tt>valid?</tt>.
       #
-      #   class Book < ActiveRecord::Base
-      #     has_many :pages
-      #     belongs_to :library
+      #   class Book
+      #     include ActiveModel::Validations
       #
-      #     validates_associated :pages, :library
+      #     attr_accessor :author, :title
+      #
+      #     validates_model :author
       #   end
       #
       # WARNING: This validation must not be used on both ends of an association.
@@ -31,11 +44,6 @@ module ActiveRecord
       # Configuration options:
       #
       # * <tt>:message</tt> - A custom error message (default is: "is invalid").
-      # * <tt>:on</tt> - Specifies the contexts where this validation is active.
-      #   Runs in all validation contexts by default +nil+. You can pass a symbol
-      #   or an array of symbols. (e.g. <tt>on: :create</tt> or
-      #   <tt>on: :custom_validation_context</tt> or
-      #   <tt>on: [:create, :custom_validation_context]</tt>)
       # * <tt>:if</tt> - Specifies a method, proc, or string to call to determine
       #   if the validation should occur (e.g. <tt>if: :allow_validation</tt>,
       #   or <tt>if: Proc.new { |user| user.signup_step > 2 }</tt>). The method,
@@ -45,8 +53,8 @@ module ActiveRecord
       #   or <tt>unless: Proc.new { |user| user.signup_step <= 2 }</tt>). The
       #   method, proc, or string should return or evaluate to a +true+ or +false+
       #   value.
-      def validates_associated(*attr_names)
-        validates_with AssociatedValidator, _merge_attributes(attr_names)
+      def validates_model(*attr_names)
+        validates_with ModelValidator, _merge_attributes(attr_names)
       end
     end
   end