+++ activemodel/CHANGELOG.md
@@ -1,3 +1,36 @@
+*   Add `ActiveModel::Validations::ModelValidator`
+
+    This validator is similar to `ActiveRecord::Validations::AssociatedValidator`
+    but doesn't require the associated object to be an `ActiveRecord` object.
+
+    It is useful when building form objects that aren't backed by a database or
+    to save multiple objects at once.
+
+    Ex:
+
+        class Author
+          include ActiveModel::Model
+
+          validates_presence_of :name
+        end
+
+        class Book
+          include ActiveModel::Model
+
+          attr_accessor :author, :title
+
+          validates_model :author
+        end
+
+        author = Author.new
+        book = Book.new(title: "A book", author: author)
+        book.valid? # => false
+
+        book.errors[:author] # => ["is invalid"]
+        author.errors[:name] # => ["can't be blank"]
+
+    *Matheus Richard*
+
 *   Add a load hook `active_model_translation` for `ActiveModel::Translation`.
 
     *Shouichi Kamiya*